Move back the components
