+++ apps/web/src/app/tools/import-export/import-export-routing.module.ts
@@ -20,4 +20,4 @@ const routes: Routes = [
 @NgModule({
   imports: [RouterModule.forChild(routes)],
 })
-export class ImportersRoutingModule {}
+export class ImportExportRoutingModule {}
